refactor: replace old launch files with dual_pose_graph.launch.py
Remove semantic_slam_launch.py, rs_semantic_slam_launch.py, and
semantic_slam_parameters_launch.py. Add clean dual_pose_graph.launch.py
with use_sim_time defaulting to false.
