--- dual_pose_graph/launch/semantic_slam_launch.py
@@ -1,71 +1,0 @@
-# Copyright 2024 Universidad Politécnica de Madrid
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#    * Redistributions of source code must retain the above copyright
-#      notice, this list of conditions and the following disclaimer.
-#
-#    * Redistributions in binary form must reproduce the above copyright
-#      notice, this list of conditions and the following disclaimer in the
-#      documentation and/or other materials provided with the distribution.
-#
-#    * Neither the name of the Universidad Politécnica de Madrid nor the names of its
-#      contributors may be used to endorse or promote products derived from
-#      this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-"""Launch file for semantic SLAM."""
-
-import os
-from launch_ros.actions import Node
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument
-from launch.substitutions import LaunchConfiguration
-from as2_core.declare_launch_arguments_from_config_file import DeclareLaunchArgumentsFromConfigFile
-from as2_core.launch_configuration_from_config_file import LaunchConfigurationFromConfigFile
-from ament_index_python.packages import get_package_share_directory
-
-
-def generate_launch_description():
-    package_folder = get_package_share_directory(
-        'dual_pose_graph')
-    config_file = os.path.join(package_folder, 'config', 'config.yaml')
-
-    """Launch dual pose graph SLAM node."""
-    return LaunchDescription([
-        DeclareLaunchArgument('namespace',
-                              description='Drone namespace',
-                              default_value='drone'),
-
-        DeclareLaunchArgument('use_sim_time',
-                              description='Use simulation clock',
-                              default_value='True'),
-
-        DeclareLaunchArgumentsFromConfigFile('config_file',
-                                             config_file,
-                                             description='Configuration file'),
-        Node(
-            package='dual_pose_graph',
-            namespace=LaunchConfiguration('namespace'),
-            executable='dual_pose_graph_node',
-            name='dual_pose_graph_node',
-            output='screen',
-            parameters=[{'use_sim_time': LaunchConfiguration('use_sim_time')},
-                        LaunchConfigurationFromConfigFile(
-                            'config_file', default_file=config_file),
-                        ],
-            emulate_tty=True,
-        ),
-    ])
